requirements.txt
robotframework==7.0.1
selenium==3.141.0
robotframework-seleniumlibrary==3.1.1
python-dotenv==0.20.0
robotframework-jsonlibrary==0.5
